reworking host to only act as realy. tools need to be adjusted now

+++ ClusterRaymarcherBase/src/usb/usb_serial.h
@@ -47,7 +47,7 @@ public:
             ;
     }
 
-    inline uint8_t available(void)
+    inline size_t available(void)
     {
         return USBSerial_available();
     }
